Moved staking mutations to the mutations directory individually
