+++ src/utils/progressNotification.ts
@@ -1,0 +1,28 @@
+import {
+  ReactNotificationOptions,
+  store as NotificationManager,
+} from "react-notifications-component";
+
+export async function usingProgressNotification<T>(
+  title: string,
+  message: string,
+  notificationType: ReactNotificationOptions["type"],
+  promise: Promise<T>
+): Promise<T> {
+  const notification = NotificationManager.addNotification({
+    container: "bottom-center",
+    title: title,
+    dismiss: {
+      click: false,
+      touch: false,
+      duration: 0,
+    },
+    message: message,
+    type: notificationType,
+  });
+  try {
+    return await promise;
+  } finally {
+    NotificationManager.removeNotification(notification);
+  }
+}
